always issue an administrative open (LCP) when PPPoED starts a new session
Hopefully this fixes a bug that cannot be reproduced outside of production: LCP sometimes remains closed after a line resync. This should ensure that there is no way LCP can get stuck in the closed state even though it doesn't address the unknown root cause. Only time and testing in production will tell if this works. Check the commit history for a revert to find out.